move job and device wrappers to providers module
