+++ tests/providers/test_braket_device.py
@@ -20,9 +20,9 @@ from braket.aws.aws_device import AwsDevice
 from braket.tasks.quantum_task import QuantumTask as AwsQuantumTask
 from qiskit import QuantumCircuit as QiskitCircuit
 
-from qbraid import QbraidError, device_wrapper
+from qbraid import QbraidError
 from qbraid.interface import random_circuit
-from qbraid.providers import QuantumJob
+from qbraid.providers import QuantumJob, device_wrapper
 from qbraid.providers.aws import BraketDevice, BraketQuantumTask
 from qbraid.providers.exceptions import ProgramValidationError
