+++ tests/providers/test_qiskit_provider.py
@@ -21,7 +21,7 @@ from qiskit_ibm_provider import IBMProvider
 from qiskit_ibm_provider.job.ibm_circuit_job import IBMCircuitJob
 from qiskit_ibm_runtime import IBMBackend, QiskitRuntimeService
 
-from qbraid import device_wrapper
+from qbraid.providers import device_wrapper
 from qbraid.providers.exceptions import JobError
 from qbraid.providers.ibm import QiskitProvider, QiskitRuntime
